feat: migrate render_utils.spec
